Added the authorization for posts so you can only edit and delelte your own posts. The only thing left is to be sure that I fix all the test cases.

+++ lib/blog_web/controllers/post_controller.ex
@@ -27,6 +27,23 @@ defmodule BlogWeb.PostController do
     render(conn, :index, posts: posts)
   end
 
+  @spec search(Plug.Conn.t(), any()) :: Plug.Conn.t()
+  def search(conn, %{"title" => title}) do
+    matching_posts = Posts.search_posts(title)
+
+    render(conn, :search_results, post: matching_posts)
+  end
+
+  def search(conn, _param) do
+    render(conn, :search_form)
+  end
+
+  @spec show(Plug.Conn.t(), map()) :: Plug.Conn.t()
+  def show(conn, %{"id" => id}) do
+    post = Posts.get_post!(id)
+    render(conn, :show, post: post)
+  end
+
   @spec new(Plug.Conn.t(), any()) :: Plug.Conn.t()
   def new(conn, _params) do
     changeset = Posts.change_post(%Post{})
@@ -46,12 +63,6 @@ defmodule BlogWeb.PostController do
       {:error, %Ecto.Changeset{} = changeset} ->
         render(conn, :new, changeset: changeset)
     end
-  end
-
-  @spec show(Plug.Conn.t(), map()) :: Plug.Conn.t()
-  def show(conn, %{"id" => id}) do
-    post = Posts.get_post!(id)
-    render(conn, :show, post: post)
   end
 
   @spec edit(Plug.Conn.t(), map()) :: Plug.Conn.t()
@@ -83,39 +94,36 @@ defmodule BlogWeb.PostController do
     end
   end
 
-  @spec update(Plug.Conn.t(), map()) :: Plug.Conn.t()
-  def update(conn, %{"id" => id, "post" => post_params}) do
-    post = Posts.get_post!(id)
+  # @spec update(Plug.Conn.t(), map()) :: Plug.Conn.t()
+  # def update(conn, %{"id" => id, "post" => post_params}) do
+  #   post = Posts.get_post!(id)
 
-    case Posts.update_post(post, post_params) do
-      {:ok, post} ->
-        conn
-        |> put_flash(:info, "Post updated successfully.")
-        |> redirect(to: ~p"/posts/#{post}")
+  #   case Posts.update_post(post, post_params) do
+  #     {:ok, post} ->
+  #       conn
+  #       |> put_flash(:info, "Post updated successfully.")
+  #       |> redirect(to: ~p"/posts/#{post}")
 
-      {:error, %Ecto.Changeset{} = changeset} ->
-        render(conn, :edit, post: post, changeset: changeset)
-    end
-  end
+  #     {:error, %Ecto.Changeset{} = changeset} ->
+  #       render(conn, :edit, post: post, changeset: changeset)
+  #   end
+  # end
 
   @spec delete(Plug.Conn.t(), map()) :: Plug.Conn.t()
   def delete(conn, %{"id" => id}) do
     post = Posts.get_post!(id)
-    {:ok, _post} = Posts.delete_post(post)
 
-    conn
-    |> put_flash(:info, "Post deleted successfully.")
-    |> redirect(to: ~p"/posts")
-  end
+    if conn.assigns[:current_user].id == post.user_id do
+      {:ok, _post} = Posts.delete_post(post)
 
-  @spec search(Plug.Conn.t(), any()) :: Plug.Conn.t()
-  def search(conn, %{"title" => title}) do
-    matching_posts = Posts.search_posts(title)
-
-    render(conn, :search_results, post: matching_posts)
-  end
-
-  def search(conn, _param) do
-    render(conn, :search_form)
+      conn
+      |> put_flash(:info, "Post deleted successfully.")
+      |> redirect(to: ~p"/posts")
+    else
+      conn
+      |> put_flash(:error, "You can only delete your own posts.")
+      |> redirect(to: ~p"/posts/#{id}")
+      |> halt()
+    end
   end
 end
